Update readme. Move CreateTrainingConfig to RunTraining.py
