+++ processor/tasks/CreateTrainingDatasets.py
@@ -283,240 +283,3 @@ class CreateTrainingDataShard(HTCondorWorkflow, law.LocalWorkflow):
         console.log("File copy out end.")
         # Remove data directory
         rmtree(prefix)
-
-
-# Task to create config files for the NN trainings
-# One config file is created for each valid combination of:
-# era, channel, mass and batch
-# For this the shards are combined and reused as necessary
-class CreateTrainingConfig(Task):
-    # Define luigi parameters
-    era = luigi.Parameter(description="Run era")
-    channel = luigi.Parameter(description="Decay Channel")
-    masses = luigi.ListParameter(
-        description="List of mass hypotheses of heavy NMSSM Higgs boson."
-    )
-    batch_nums = luigi.ListParameter(
-        description="List of groups of mass hypotheses of light NMSSM Higgs boson."
-    )
-    processes_and_classes = luigi.ListParameter(
-        description="List of processes and their classes used in training"
-    )
-
-    all_eras = ["2016", "2017", "2018"]
-
-    # Set other variables and templates used by this class.
-    dir_template = "{era}_{channel}_{mass}_{batch}"
-    dir_template_in = "{era}_{channel}"
-    file_template = "training_config.yaml"
-    files_template_shard = "{process}_datashard_fold{fold}.root"
-    files_template_shard_config = "{process}_datashard_config.yaml"
-
-    # Function to check for invalid mass-batch combinations
-    def valid_batches(self, mass):
-        mass = str(mass)
-        if mass in ["240", "280"]:
-            max_batch = 3
-        elif mass in ["320", "360", "400", "450"]:
-            max_batch = 4
-        elif mass in ["500", "550", "600"]:
-            max_batch = 5
-        elif mass in ["700", "800", "heavier"]:
-            max_batch = 6
-        elif mass in ["900", "1000"]:
-            max_batch = 7
-        else:
-            raise Exception("Provided mass {} is not valid.".format(mass))
-        valid_batches = [batch for batch in self.batch_nums if int(batch) <= max_batch]
-        return valid_batches
-
-    # Set prerequisites of this task:
-    # All configs for the dataset shards have to be completed
-    # All dataset shards have to be completed
-    def requires(self):
-        if self.era == "all_eras":
-            use_eras = self.all_eras
-        else:
-            use_eras = [self.era]
-
-        requirements = {}
-        for era in use_eras:
-            requirements_args = {
-                "era": era,
-                "channel": self.channel,
-                "processes_and_classes": self.processes_and_classes,
-                "identifier": [self.channel],
-            }
-            requirements["CreateTrainingDataShard_{}".format(era)] = PuppetMaster(
-                puppet_task=CreateTrainingDataShard(**requirements_args),
-                identifier=[era, self.channel],
-            )
-            requirements["CreateTrainingDataShardConfig_{}".format(era)] = PuppetMaster(
-                puppet_task=CreateTrainingDataShardConfig(**requirements_args),
-                identifier=[era, self.channel],
-            )
-        return requirements
-
-    # Define output targets. Task is considerd complete if all targets are present.
-    def output(self):
-        files = [
-            "/".join(
-                [
-                    self.dir_template.format(
-                        era=self.era,
-                        channel=self.channel,
-                        mass=mass,
-                        batch=batch_num,
-                    ),
-                    self.file_template,
-                ]
-            )
-            for mass in self.masses
-            for batch_num in self.valid_batches(mass)
-        ]
-        targets = self.remote_targets(files)
-        for target in targets:
-            target.parent.touch()
-        return targets
-
-    def run(self):
-        if self.era == "all_eras":
-            use_eras = self.all_eras
-        else:
-            use_eras = [self.era]
-        processes, process_classes = zip(*self.processes_and_classes)
-        prefix = self.temporary_local_path("")
-        run_loc = "sm-htt-analysis"
-        # Create data directory for each valid branch and mass/batch combination
-        for era in use_eras + [self.era]:
-            for mass in self.masses:
-                for batch_num in self.valid_batches(mass):
-                    os.makedirs(
-                        "/".join(
-                            [
-                                prefix,
-                                self.dir_template.format(
-                                    era=era,
-                                    channel=self.channel,
-                                    mass=mass,
-                                    batch=batch_num,
-                                ),
-                            ]
-                        ),
-                        exist_ok=True,
-                    )
-
-        files = [
-            "/".join(
-                [
-                    prefix,
-                    self.dir_template.format(
-                        era=self.era,
-                        channel=self.channel,
-                        mass=mass,
-                        batch=batch_num,
-                    ),
-                    self.file_template,
-                ]
-            )
-            for mass in self.masses
-            for batch_num in self.valid_batches(mass)
-        ]
-
-        # Get list of all shard targets
-        allbranch_shards = {
-            era: flatten_collections(
-                self.requires()[
-                    "CreateTrainingDataShard_{}".format(era)
-                ].give_puppet_outputs()["collection"]
-            )
-            for era in use_eras
-        }
-        # Get prefix of remote storage for root shards
-        remote_shard_base = {
-            era: self.wlcg_path + os.path.dirname(allbranch_shards[era][0].path)
-            for era in use_eras
-        }
-
-        # Get shard config targets
-        branch_shardconfigs = {
-            era: self.requires()["CreateTrainingDataShardConfig_{}".format(era)]
-            .give_puppet_outputs()
-            .targets
-            for era in use_eras
-        }
-
-        for era in use_eras:
-            assert len(branch_shardconfigs[era]) == len(
-                processes
-            ), "There should be {} targets, but there are {}".format(
-                len(processes),
-                len(branch_shardconfigs[era]),
-            )
-
-        # Copy shard config files into data directory
-        console.log("File copy in start.")
-        for era in use_eras:
-            for copy_file in branch_shardconfigs[era]:
-                copy_file_name = os.path.basename(copy_file.path)
-                copy_file_name_path = os.path.basename(os.path.dirname(copy_file.path))
-                copy_file.copy_to_local(
-                    "/".join([prefix, copy_file_name_path, copy_file_name])
-                )
-        console.log("File copy in end.")
-        # Write training config file
-        for era in use_eras:
-            self.run_command(
-                [
-                    "python",
-                    "ml_datasets/write_training_config.py",
-                    "--era {}".format(era),
-                    "--channel {}".format(self.channel),
-                    "--masses {}".format(" ".join(self.masses)),
-                    "--batches {}".format(" ".join(self.batch_nums)),
-                    "--config-dir {}/{}_{}".format(prefix, era, self.channel),
-                    "--dataset-dir {}".format(remote_shard_base[era]),
-                    "--processes {}".format(" ".join(processes)),
-                    "--training-template datasets/templates/{}_{}_training.yaml".format(
-                        era, self.channel
-                    ),
-                    "--write-weights True",
-                    "--overwrite-configs {}".format(""),
-                ],
-                run_location=run_loc,
-                sourcescripts="{}/utils/setup_python.sh".format(run_loc),
-            )
-        if self.era == "all_eras":
-            # Create data directory
-            for mass in self.masses:
-                for batch_num in self.valid_batches(mass):
-                    conf_files = "/".join(
-                        [
-                            prefix,
-                            self.dir_template.format(
-                                era="{era}",
-                                channel=self.channel,
-                                mass=mass,
-                                batch=batch_num,
-                            ),
-                        ]
-                    )
-                    # Combine the configs of each era
-                    self.run_command(
-                        [
-                            "python",
-                            "ml_datasets/create_combined_config.py",
-                            "--input-path {}".format(conf_files),
-                            "--output-dir {}".format(conf_files.format(era="all_eras")),
-                        ],
-                        run_location=run_loc,
-                    )
-
-        # Copy locally created files to remote storage
-        console.log("File copy out start.")
-        for file_remote, file_local in zip(self.output(), files):
-            file_remote.parent.touch()
-            file_remote.copy_from_local(file_local)
-        console.log("File copy out end.")
-        # Remove data directories
-        rmtree(prefix)
